DATA: Migrate service schemas from NUMERIC IDs to BIGINT

+++ services/auth/db_setup.sql
@@ -1,5 +1,5 @@
 CREATE TABLE app_user(
-    user_id NUMERIC(22, 0) PRIMARY KEY,
+    user_id BIGINT PRIMARY KEY,
      username VARCHAR(50) NOT NULL UNIQUE,
      password VARCHAR(60) NOT NULL,
      email VARCHAR(100) NOT NULL UNIQUE
@@ -7,13 +7,13 @@ CREATE TABLE app_user(
 
 CREATE TABLE refresh_tokens (
     token_id   TEXT PRIMARY KEY,
-    user_id    NUMERIC(22, 0) NOT NULL,
+    user_id    BIGINT NOT NULL,
     expires_at TIMESTAMPTZ NOT NULL,
     revoked    BOOLEAN NOT NULL DEFAULT false,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
 
 CREATE TABLE user_token_versions (
-    user_id NUMERIC(22, 0) PRIMARY KEY,
+    user_id BIGINT PRIMARY KEY,
     version INTEGER NOT NULL DEFAULT 1
 );
