+++ services/restaurant/db_setup.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS restaurants (
-    id NUMERIC(22, 0) PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     name VARCHAR(100) UNIQUE NOT NULL,
     latitude DOUBLE PRECISION NOT NULL,
     longitude DOUBLE PRECISION NOT NULL,
@@ -15,7 +15,7 @@ CREATE TABLE IF NOT EXISTS restaurants (
 
 CREATE TABLE IF NOT EXISTS restaurant_categories (
     id SERIAL PRIMARY KEY,
-    restaurant_id NUMERIC(22, 0) NOT NULL REFERENCES restaurants(id) ON DELETE CASCADE,
+    restaurant_id BIGINT NOT NULL REFERENCES restaurants(id) ON DELETE CASCADE,
     category TEXT NOT NULL
 );
